Fix Vulkan backend issues for Windows/RTX 50 series
1. NRIDescs.h + DeviceVK.hpp: Use int32_t for viewport bounds
   - RTX 5090 reports viewportBoundsRange of [-65536, 65536]
   - Changed boundsMin/boundsMax from int16_t to int32_t to support full range

2. CommandBufferVK.hpp: Fix null pointer dereference in texture barriers
   - srcQueue/dstQueue are optional but were dereferenced to create references
     before checking for null (undefined behavior)
   - Changed from references to pointers: & → *, . → ->

+++ Include/NRIDescs.h
@@ -1729,8 +1729,8 @@ NriStruct(DeviceDesc) {
     // Viewport
     struct {
         uint32_t maxNum;
-        int16_t boundsMin;
-        int16_t boundsMax;
+        int32_t boundsMin;
+        int32_t boundsMax;
     } viewport;
 
     // Dimensions
